Add spesific cases for the different widgets that can open
Not done with all of them but you can now use a mouse when inventory opens.

+++ SandboxS24/Source/SandboxS24/Public/HUD_BaseSetup.h
@@ -53,7 +53,7 @@ public:
 	//Modals (Dialogs, Popups, etc.)
 
 public:
-	float CurrentID;
+	int CurrentID;
 	void CollapseAllWidgets();
 	void WantsToOpenWidget(int WidgetID);
 	void ExpandWidget(int WidgetID);
